Remove orphaned dual descriptive hover panels after surface migration.
Policy-owned composition on the map interaction surface is the only path; keeping blockedByPlanetHover panels invited a second architecture.
