--- packages/frontend/src/analytics/stellar-cartography/StellarCartographyHoverPanel.tsx
@@ -1,153 +1,0 @@
-import { useEffect, useRef, useState } from 'react'
-import { useStore } from '@xyflow/react'
-import {
-  fetchStellarCartographySample,
-  type StellarCartographySampleEntry,
-} from '../../api/bff'
-import type { StellarCartographyMapContext } from './mapUiConfig'
-import type { CartographyVisibilityPolicy } from './cartographyVisibilityPolicy'
-import { flowToMapCellIndices } from '../../lib/planetSpatialGrid'
-import { formatStellarCartographySampleLine } from './sampleTooltipFormat'
-
-const SAMPLE_DEBOUNCE_MS = 100
-
-type PaneClientPos = { x: number; y: number }
-
-type StellarCartographyHoverPanelProps = {
-  cartography: StellarCartographyMapContext
-  wormholeHoverLines: string[] | null
-  /** Shared pane pointer from ``useMapPaneClientPos`` (no listener of its own). */
-  clientPos: PaneClientPos | null
-  /** When a planet hover/pin label is showing, suppress cartography hover entirely. */
-  blockedByPlanetHover?: boolean
-  /** Extra stacked lines (e.g. homeworld region hover copy). */
-  additionalHoverLines?: readonly string[] | null
-  clientToFlowPosition: (
-    clientX: number,
-    clientY: number,
-    domNode: HTMLElement | null,
-    transform: [number, number, number] | undefined,
-    paneRect?: Pick<DOMRect, 'left' | 'top'>
-  ) => { x: number; y: number } | null
-}
-
-/** Build stacked hover lines for all active cartography features at a map cell. */
-export function buildStellarCartographyHoverLines(
-  entries: StellarCartographySampleEntry[],
-  wormholeHoverLines: string[] | null,
-  policy: CartographyVisibilityPolicy,
-  additionalHoverLines: readonly string[] | null = null
-): string[] {
-  const lines = policy.sampleEntries(entries).map(formatStellarCartographySampleLine)
-  if (
-    wormholeHoverLines != null &&
-    wormholeHoverLines.length > 0 &&
-    policy.areWormholesShown()
-  ) {
-    lines.push(...wormholeHoverLines)
-  }
-  if (additionalHoverLines != null && additionalHoverLines.length > 0) {
-    lines.push(...additionalHoverLines)
-  }
-  return lines
-}
-
-/** Debounced stacked tooltip for Stellar Cartography map hover sampling (Phase 4b). */
-export function StellarCartographyHoverPanel({
-  cartography,
-  wormholeHoverLines,
-  clientPos,
-  blockedByPlanetHover = false,
-  additionalHoverLines = null,
-  clientToFlowPosition,
-}: StellarCartographyHoverPanelProps) {
-  const domNode = useStore((s) => s.domNode ?? null)
-  const transform = useStore((s) => s.transform)
-  const [entries, setEntries] = useState<StellarCartographySampleEntry[]>([])
-  const debounceRef = useRef<ReturnType<typeof setTimeout> | null>(null)
-  const requestSeqRef = useRef(0)
-
-  useEffect(() => {
-    if (debounceRef.current != null) {
-      clearTimeout(debounceRef.current)
-      debounceRef.current = null
-    }
-    if (clientPos == null || !transform) {
-      setEntries([])
-      return
-    }
-    if (blockedByPlanetHover) {
-      setEntries([])
-      return
-    }
-    const flow = clientToFlowPosition(clientPos.x, clientPos.y, domNode, transform)
-    if (flow == null) {
-      setEntries([])
-      return
-    }
-    const { mapX, mapY } = flowToMapCellIndices(flow.x, flow.y)
-    const seq = ++requestSeqRef.current
-    debounceRef.current = setTimeout(() => {
-      debounceRef.current = null
-      void fetchStellarCartographySample(cartography.analyticScope, mapX, mapY)
-        .then((data) => {
-          if (seq !== requestSeqRef.current) return
-          setEntries(data.entries)
-        })
-        .catch(() => {
-          if (seq !== requestSeqRef.current) return
-          setEntries([])
-        })
-    }, SAMPLE_DEBOUNCE_MS)
-    return () => {
-      if (debounceRef.current != null) {
-        clearTimeout(debounceRef.current)
-        debounceRef.current = null
-      }
-    }
-  }, [
-    cartography,
-    clientPos,
-    clientToFlowPosition,
-    domNode,
-    blockedByPlanetHover,
-    transform,
-  ])
-
-  const lines = buildStellarCartographyHoverLines(
-    entries,
-    wormholeHoverLines,
-    cartography.policy,
-    additionalHoverLines
-  )
-  if (
-    blockedByPlanetHover ||
-    lines.length === 0 ||
-    clientPos == null ||
-    domNode == null
-  ) {
-    return null
-  }
-
-  const rect = domNode.getBoundingClientRect()
-  const paneX = clientPos.x - rect.left + 12
-  const paneY = clientPos.y - rect.top - 8
-
-  return (
-    <div
-      className="pointer-events-none absolute z-[6] max-w-xs font-mono text-xs text-gray-300"
-      style={{
-        left: paneX,
-        top: paneY,
-        transform: 'translateY(-100%)',
-        backgroundColor: '#000000',
-        borderRadius: 6,
-        padding: '4px 8px',
-      }}
-    >
-      {lines.map((line, i) => (
-        <div key={`${i}-${line}`}>{line}</div>
-      ))}
-    </div>
-  )
-}
